arreglo perfecto hecho para los dioses. Un buen Oier en su prime

+++ src/main/java/eus/fpsanturtzilh/repositories/MaterialakRepository.java
@@ -3,7 +3,7 @@ package eus.fpsanturtzilh.repositories;
 import eus.fpsanturtzilh.models.Materialak;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialakRepository extends JpaRepository<Materialak, Long> {
+public interface MaterialakRepository extends JpaRepository<Materialak, Integer> {
 	
 }
