+++ src/main/java/eus/fpsanturtzilh/repositories/ProduktuakRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProduktuakRepository extends JpaRepository<Produktuak, Integer> {
+public interface ProduktuakRepository extends JpaRepository<Produktuak, Long> {
+	
 }
